requirements.txt
rembg==2.0.66
Pillow==10.2.0
gunicorn==20.1.0
Flask==2.3.3
Werkzeug==2.3.7
onnxruntime==1.15.0
numpy==1.24.4
boto3
python-dotenv
mysql-connector-python
warrant
zappa